+++ cmd/gitcrawl/signals_windows.go
@@ -1,0 +1,9 @@
+//go:build windows
+
+package main
+
+import "os"
+
+func commandSignals() []os.Signal {
+	return []os.Signal{os.Interrupt}
+}